requirements.txt
boto3>=1.34.0
gql[all]>=3.5.0
requests-aws4auth>=1.2.3
openpyxl>=3.1.2
python-dotenv>=1.0.0

